Huge Improvements
Completely changed how the main menu worked, Added a file that will be created after using this debloated to check if the folder you're trying to debloat has been debloated in the past and if yes it will block the functions from triggering. Now you get asked to  remove the Debloated {Folder} that has been created in the previous debloat operation.